requirements.txt
aiohttp==3.6.0
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Django==2.2.5
djangorestframework==3.10.3
Flask==1.1.1
get==2019.4.13
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.5.2
post==2019.4.13
public==2019.4.13
pyee==6.0.0
pytz==2019.2
query-string==2019.4.13
redis==3.3.8
request==2019.4.13
requests==2.22.0
six==1.12.0
slackclient==2.1.0
slackeventsapi==2.1.0
sqlparse==0.3.0
typing-extensions==3.7.4
urllib3==1.25.3
Werkzeug==0.15.6
yarl==1.3.0
